Fixes a deprecation in Rails 6 and error in Rails 7.
The following deprecation notice was shown in Rails 6:

DEPRECATION WARNING: Single arity template handlers are deprecated. Template handlers must
now accept two parameters, the view object and the source for the view object.
Change:
  >> Prawn::Rails::TemplateHandler.call(template)
To:
  >> Prawn::Rails::TemplateHandler.call(template, source)

This warning became an error in Rails 7, and as it appears that this project is abandoned we are making our own fork and implementing this simple fix.